Bug 1460026 - separate an "api" from a "service"
APIs are declared, and the setup function returns a service object that
is specialized to the given rootUrl.

This is a big refactor with some JS modernization along the way.  It
also upgrades lib-api to be compatible with other concurrent library
changes such as lib-validate requiring a rootUrl and serviceName.