Add Cache-Control to the Access-Control-Allow-Headers header (#107)
The new implementation of tc-events uses an EventSource to
set up communication between the client and server. Attempting
to set up a connection throws an error in the console:
`Request header field Cache-Control is not allowed by
Access-Control-Allow-Headers in preflight response.`.

I'm not sure you can set the headers in the EventSource
interface from the client without a polyfill which is non-standard.
That being said, this commit allows Cache-Control to be sent from the
client.